requirements.txt
autopep8==2.1.0
black==24.3.0
flake8==7.0.0
iniconfig==2.0.0
pathspec==0.12.1
pytest==8.1.1
tomli==2.0.1
redis==5.0.4
Django==5.0.6
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
djangorestframework==3.15.1

